+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v5.0.0-1.19.3] - 2023-03-20
+## [v5.0.0-1.19.3] - 2023-03-21
 - Ported to Minecraft 1.19.3
+### Added
+- Enchanting table supports chiseled bookshelves, for every 3 contained books one level of enchanting power is added to the table (note that the chiseled bookshelf must be facing the enchanting table)
+- Chiseled bookshelves show a tooltip for the book you are looking at
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/